Add files via upload
Tower of Hanoi, is a mathematical puzzle which consists of three towers (pegs) and more than one rings.
Rules : 
The mission is to move all the disks to some another tower without violating the sequence of arrangement. A few rules to be followed for Tower of Hanoi are −
Only one disk can be moved among the towers at any given time.
Only the "top" disk can be removed.
No large disk can sit over a small disk.